Visualize XY Fields
Added function for visualizing XY fields.  This has not been tested, but
just a copy past with transforming the field index locations to match an
X, Y, t slice of data.  For now it only outputs the fields in the
y-direction.

Need to update the visualization functions to output all fields.